implemented authorized requests

+++ src/client/src/components/PostDetails.js
@@ -1,6 +1,7 @@
 // import { useEffect } from 'react'
 import { usePostsContext } from '../hooks/usePostsContext'
 import { Link } from 'react-router-dom'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
@@ -10,10 +11,17 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const PostDetails = ({ post }) => {
 
     const { dispatch } = usePostsContext()
+    const { user } = useAuthContext
 
     const handleDelete = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/posts/' + post._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization' : `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
